Changed data collection stuff to 64-bit values.

+++ watchlib/auto_wp.h
@@ -36,16 +36,16 @@ namespace Hongyi_WatchPoint {
 	};
 
 	struct trie_data_t {
-		unsigned int top_hit;
-		unsigned int mid_hit;
-		unsigned int bot_hit;
+		unsigned long long top_hit;
+		unsigned long long mid_hit;
+		unsigned long long bot_hit;
 		
-		unsigned int top_change;
-		unsigned int mid_change;
-		unsigned int bot_change;
+		unsigned long long top_change;
+		unsigned long long mid_change;
+		unsigned long long bot_change;
 		
-		unsigned int top_break;
-		unsigned int mid_break;
+		unsigned long long top_break;
+		unsigned long long mid_break;
 		
 		const trie_data_t operator+(const trie_data_t &other) const;
 		trie_data_t();
@@ -123,7 +123,7 @@ namespace Hongyi_WatchPoint {
 		//all below are private
 		void rm_watchpoint	(ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags);
 		void add_watchpoint	(ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags);
-		bool general_fault	(ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags,  unsigned int& top_page, unsigned int& mid_page, unsigned int& bot_page);
+		bool general_fault	(ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags,  unsigned long long& top_page, unsigned long long& mid_page, unsigned long long& bot_page);
 		PAGE_HIT page_level	(ADDRESS target_addr, typename deque<watchpoint_t<ADDRESS, FLAGS> >::iterator iter);
 		void page_break		(PAGE_HIT& before, PAGE_HIT& after);
 		
@@ -1075,7 +1075,7 @@ namespace Hongyi_WatchPoint{
 	}
 	
 	template <class ADDRESS, class FLAGS>
-	bool WatchPoint<ADDRESS, FLAGS>::general_fault (ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags, unsigned int& top_page, unsigned int& mid_page, unsigned int& bot_page) {
+	bool WatchPoint<ADDRESS, FLAGS>::general_fault (ADDRESS target_addr, ADDRESS target_size, FLAGS target_flags, unsigned long long& top_page, unsigned long long& mid_page, unsigned long long& bot_page) {
 		//	cout << "Entered general_fault" << endl;
 
 		if (target_size == 0) {
